+++ src/lightning_app/cli/lightning_cli.py
@@ -358,8 +358,8 @@ def run_app(
     )
 
 
-if RequirementCache("lightning-fabric"):
-    # lightning-fabric may not be available when installing only standalone lightning-app package
+if RequirementCache("lightning-fabric>=1.9.0.dev0") or RequirementCache("lightning>=1.9.0.dev0"):
+    # lightning.fabric.cli may not be available when installing only standalone lightning-app package
     from lightning_fabric.cli import _run_model
 
     run.add_command(_run_model)