+++ tests/tests_app/core/scripts/lightning_overrides.py
@@ -9,10 +9,10 @@ if _is_pytorch_lightning_available():
     from pytorch_lightning import LightningDataModule, LightningModule, Trainer
     from pytorch_lightning.accelerators.accelerator import Accelerator
     from pytorch_lightning.callbacks import Callback
-    from pytorch_lightning.loggers import LightningLoggerBase
-    from pytorch_lightning.loops.base import Loop
+    from pytorch_lightning.loggers import Logger
+    from pytorch_lightning.loops import Loop
     from pytorch_lightning.plugins import PrecisionPlugin
-    from pytorch_lightning.profiler.base import BaseProfiler
+    from pytorch_lightning.profilers import Profiler
 
 
 if __name__ == "__main__":
@@ -38,7 +38,7 @@ if __name__ == "__main__":
     class BoringCallback(Callback):
         pass
 
-    class BoringLightningLoggerBase(LightningLoggerBase):
+    class BoringLogger(Logger):
         pass
 
     class BoringLoop(Loop):
@@ -47,5 +47,5 @@ if __name__ == "__main__":
     class BoringMetric(Metric):
         pass
 
-    class BoringBaseProfiler(BaseProfiler):
+    class BoringProfiler(Profiler):
         pass